requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==11.1.0
opencv-python==4.10.0.84
python-dotenv==1.0.0
Werkzeug==2.3.7
marshmallow==3.20.1
pytest==7.4.2
pytest-flask==1.2.0
requests==2.31.0
numpy==2.3.2

# LLM providers for enhanced extraction
openai==1.12.0
google-generativeai==0.3.2
anthropic==0.8.1
